+++ src/components/Navbar.js
@@ -1,4 +1,4 @@
-import './App.css';
+import '../App.css';
 import { Link } from 'react-router-dom';
 
 export default function Navbar() {
