+++ src/components/UclStandings.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useEffect } from "react";
 import { useState } from "react";
-import "./App.css";
+import "../App.css";
 import "./GroupStandings.css";
 
 export default function UclStandings() {
